requirements.txt
requests>=2.28.2
feedparser>=6.0.0
python-dotenv>=1.0.0
schedule>=1.2.0
openai>=0.27.0
newsapi-python>=0.2.7
nltk>=3.8.1
sumy>=0.11.0
bs4>=0.0.1
beautifulsoup4>=4.12.0
newspaper3k>=0.2.8 